+++ app/ante/vesting.go
@@ -1,9 +1,11 @@
 package ante
 
 import (
+	"github.com/cosmos/cosmos-sdk/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
+	"github.com/cosmos/cosmos-sdk/x/authz"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	evmtypes "github.com/tharsis/ethermint/x/evm/types"
 	vestingtypes "github.com/tharsis/evmos/x/vesting/types"
@@ -77,15 +79,17 @@ func (vtd EthVestingTransactionDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx,
 
 // VestingDelegationDecorator validates delegation of vested coins
 type VestingDelegationDecorator struct {
-	ak evmtypes.AccountKeeper
-	sk vestingtypes.StakingKeeper
+	ak  evmtypes.AccountKeeper
+	sk  vestingtypes.StakingKeeper
+	cdc codec.BinaryCodec
 }
 
 // NewVestingDelegationDecorator creates a new VestingDelegationDecorator
-func NewVestingDelegationDecorator(ak evmtypes.AccountKeeper, sk vestingtypes.StakingKeeper) VestingDelegationDecorator {
+func NewVestingDelegationDecorator(ak evmtypes.AccountKeeper, sk vestingtypes.StakingKeeper, cdc codec.BinaryCodec) VestingDelegationDecorator {
 	return VestingDelegationDecorator{
-		ak: ak,
-		sk: sk,
+		ak:  ak,
+		sk:  sk,
+		cdc: cdc,
 	}
 }
 
@@ -94,46 +98,79 @@ func NewVestingDelegationDecorator(ak evmtypes.AccountKeeper, sk vestingtypes.St
 // the coins already vested
 func (vdd VestingDelegationDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate bool, next sdk.AnteHandler) (newCtx sdk.Context, err error) {
 	for _, msg := range tx.GetMsgs() {
-		// Continue only if delegation
-		delegateMsg, isDelegation := msg.(*stakingtypes.MsgDelegate)
-		if !isDelegation {
-			continue
-		}
-
-		for _, addr := range msg.GetSigners() {
-			acc := vdd.ak.GetAccount(ctx, addr)
-			if acc == nil {
-				return ctx, sdkerrors.Wrapf(
-					sdkerrors.ErrUnknownAddress,
-					"account %s does not exist", addr,
-				)
+		switch msg := msg.(type) {
+		case *authz.MsgExec:
+			// Check for bypassing authorization
+			if err := vdd.validateAuthz(ctx, msg); err != nil {
+				return ctx, err
 			}
-
-			clawbackAccount, isClawback := acc.(*vestingtypes.ClawbackVestingAccount)
-			if !isClawback {
-				// continue to next decorator as this logic only applies to vesting
-				return next(ctx, tx, simulate)
-			}
-
-			// error if bond amount is > vested coins
-			bondDenom := vdd.sk.BondDenom(ctx)
-			coins := clawbackAccount.GetVestedOnly(ctx.BlockTime())
-			if coins == nil || coins.Empty() {
-				return ctx, sdkerrors.Wrap(
-					vestingtypes.ErrInsufficientVestedCoins,
-					"account has no vested coins",
-				)
-			}
-
-			vested := coins.AmountOf(bondDenom)
-			if vested.LT(delegateMsg.Amount.Amount) {
-				return ctx, sdkerrors.Wrapf(
-					vestingtypes.ErrInsufficientVestedCoins,
-					"cannot delegate unvested coins. Coins Vested: %s", vested,
-				)
+		default:
+			if err := vdd.validateMsg(ctx, msg); err != nil {
+				return ctx, err
 			}
 		}
 	}
 
 	return next(ctx, tx, simulate)
+}
+
+// validateAuthz validates the authorization internal message
+func (vdd VestingDelegationDecorator) validateAuthz(ctx sdk.Context, execMsg *authz.MsgExec) error {
+	for _, v := range execMsg.Msgs {
+		var innerMsg sdk.Msg
+		if err := vdd.cdc.UnpackAny(v, &innerMsg); err != nil {
+			return sdkerrors.Wrap(err, "cannot unmarshal authz exec msgs")
+		}
+
+		if err := vdd.validateMsg(ctx, innerMsg); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+// validateMsg checks that the only vested coins can be delegated
+func (vdd VestingDelegationDecorator) validateMsg(ctx sdk.Context, msg sdk.Msg) error {
+	delegateMsg, ok := msg.(*stakingtypes.MsgDelegate)
+	if !ok {
+		return nil
+	}
+
+	for _, addr := range msg.GetSigners() {
+		acc := vdd.ak.GetAccount(ctx, addr)
+		if acc == nil {
+			return sdkerrors.Wrapf(
+				sdkerrors.ErrUnknownAddress,
+				"account %s does not exist", addr,
+			)
+		}
+
+		clawbackAccount, isClawback := acc.(*vestingtypes.ClawbackVestingAccount)
+		if !isClawback {
+			// continue to next decorator as this logic only applies to vesting
+			return nil
+		}
+
+		// error if bond amount is > vested coins
+		bondDenom := vdd.sk.BondDenom(ctx)
+		coins := clawbackAccount.GetVestedOnly(ctx.BlockTime())
+		if coins == nil || coins.Empty() {
+			return sdkerrors.Wrap(
+				vestingtypes.ErrInsufficientVestedCoins,
+				"account has no vested coins",
+			)
+		}
+
+		vested := coins.AmountOf(bondDenom)
+		if vested.LT(delegateMsg.Amount.Amount) {
+			return sdkerrors.Wrapf(
+				vestingtypes.ErrInsufficientVestedCoins,
+				"cannot delegate unvested coins. coins vested < delegation amount (%s < %s)",
+				vested, delegateMsg.Amount.Amount,
+			)
+		}
+	}
+
+	return nil
 }