+++ x/erc20/keeper/msg_server.go
@@ -135,9 +135,10 @@ func (k Keeper) convertCoinNativeCoin(
 	tokens := msg.Coin.Amount.BigInt()
 	balanceTokenAfter := k.balanceOf(ctx, erc20, contract, receiver)
 	exp := big.NewInt(0).Add(balanceToken, tokens)
+
 	if r := balanceTokenAfter.Cmp(exp); r != 0 {
 		return nil, sdkerrors.Wrapf(
-			types.ErrInvalidConversionBalance,
+			types.ErrBalanceInvariance,
 			"invalid token balance - expected: %v, actual: %v", exp, balanceTokenAfter,
 		)
 	}
@@ -188,12 +189,13 @@ func (k Keeper) convertERC20NativeCoin(
 	if err := k.bankKeeper.SendCoinsFromModuleToAccount(ctx, types.ModuleName, receiver, coins); err != nil {
 		return nil, err
 	}
+
 	// Check expected Receiver balance after transfer execution
 	balanceCoinAfter := k.bankKeeper.GetBalance(ctx, receiver, pair.Denom)
 	expCoin := balanceCoin.Add(coins[0])
 	if ok := balanceCoinAfter.IsEqual(expCoin); !ok {
 		return nil, sdkerrors.Wrapf(
-			types.ErrInvalidConversionBalance,
+			types.ErrBalanceInvariance,
 			"invalid coin balance - expected: %v, actual: %v",
 			expCoin, balanceCoinAfter,
 		)
@@ -205,7 +207,7 @@ func (k Keeper) convertERC20NativeCoin(
 	expToken := big.NewInt(0).Sub(balanceToken, tokens)
 	if r := balanceTokenAfter.Cmp(expToken); r != 0 {
 		return nil, sdkerrors.Wrapf(
-			types.ErrInvalidConversionBalance,
+			types.ErrBalanceInvariance,
 			"invalid token balance - expected: %v, actual: %v",
 			expToken, balanceTokenAfter,
 		)
@@ -253,6 +255,7 @@ func (k Keeper) convertERC20NativeToken(
 	if err != nil {
 		return nil, err
 	}
+
 	res, err := k.CallEVMWithData(ctx, sender, &contract, transferData)
 	if err != nil {
 		return nil, err
@@ -272,9 +275,10 @@ func (k Keeper) convertERC20NativeToken(
 	tokens := coins[0].Amount.BigInt()
 	balanceTokenAfter := k.balanceOf(ctx, erc20, contract, types.ModuleAddress)
 	expToken := big.NewInt(0).Add(balanceToken, tokens)
+
 	if r := balanceTokenAfter.Cmp(expToken); r != 0 {
 		return nil, sdkerrors.Wrapf(
-			types.ErrInvalidConversionBalance,
+			types.ErrBalanceInvariance,
 			"invalid token balance - expected: %v, actual: %v",
 			expToken, balanceTokenAfter,
 		)
@@ -293,9 +297,10 @@ func (k Keeper) convertERC20NativeToken(
 	// Check expected Receiver balance after transfer execution
 	balanceCoinAfter := k.bankKeeper.GetBalance(ctx, receiver, pair.Denom)
 	expCoin := balanceCoin.Add(coins[0])
+
 	if ok := balanceCoinAfter.IsEqual(expCoin); !ok {
 		return nil, sdkerrors.Wrapf(
-			types.ErrInvalidConversionBalance,
+			types.ErrBalanceInvariance,
 			"invalid coin balance - expected: %v, actual: %v",
 			expCoin, balanceCoinAfter,
 		)
@@ -368,9 +373,10 @@ func (k Keeper) convertCoinNativeERC20(
 	tokens := msg.Coin.Amount.BigInt()
 	balanceTokenAfter := k.balanceOf(ctx, erc20, contract, receiver)
 	exp := big.NewInt(0).Add(balanceToken, tokens)
+
 	if r := balanceTokenAfter.Cmp(exp); r != 0 {
 		return nil, sdkerrors.Wrapf(
-			types.ErrInvalidConversionBalance,
+			types.ErrBalanceInvariance,
 			"invalid token balance - expected: %v, actual: %v", exp, balanceTokenAfter,
 		)
 	}
@@ -413,8 +419,8 @@ func (k Keeper) balanceOf(
 		return nil
 	}
 
-	unpacked, _ := abi.Unpack("balanceOf", res.Ret)
-	if len(unpacked) == 0 {
+	unpacked, err := abi.Unpack("balanceOf", res.Ret)
+	if err != nil || len(unpacked) == 0 {
 		return nil
 	}
 
@@ -439,9 +445,10 @@ func (k Keeper) monitorApprovalEvent(res *evmtypes.MsgEthereumTxResponse) error 
 	for _, log := range res.Logs {
 		if log.Topics[0] == logApprovalSigHash.Hex() {
 			return sdkerrors.Wrapf(
-				types.ErrUnexpectedEvent, "unexpected approval event",
+				types.ErrUnexpectedEvent, "unexpected Approval event",
 			)
 		}
 	}
+
 	return nil
 }