+++ CHANGELOG.md
@@ -35,6 +35,13 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 # Changelog
 
+## [v1.1.0] - 2022-03-02
+
+### Bug Fixes
+
+- (ante) [\#318](https://github.com/tharsis/evmos/pull/318) Add authz check in vesting and min commission `AnteHandler` decorators.  
+- (vesting) [\#317](https://github.com/tharsis/evmos/pull/317) Fix clawback for vested coins.
+
 ## [v1.0.0] - 2022-02-28
 
 ### State Machine Breaking